# docs/openapi_outline.yaml
openapi: 3.1.0
info:
  title: Agentic Material Search API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /jobs/{jobId}/artifacts/{artifactId}:
    get:
      summary: Download a generated artifact
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
        - in: path
          name: artifactId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File stream
...